Refactor: change package structure for tf.
