+++ whitelist.txt
@@ -41,6 +41,7 @@ vm
 gwei
 wei
 
+getmtime
 byteorder
 delitem
 dirname
